Comment out guide to Deploying on Heroku
Trainees expect every step by step instructions to work for all cases,
but in reality there are many ways to deploy an application. Removing
this from the guides to prevent the assumption that this is a one size
fits all guide to deploying every app.

For Sequelize, the current working steps are documented here:
https://github.com/sabrina-tw/sequelize-jwt-pokemon#heroku-deployment